move all of thread body into index
