+++ packages/commonwealth/client/scripts/views/components/component_kit/cw_content_page.tsx
@@ -6,12 +6,14 @@ import moment from 'moment';
 import 'components/component_kit/cw_content_page.scss';
 
 import { Comment } from 'models';
+import { pluralize } from 'helpers';
 import { ComponentType, MenuItem } from './types';
 import { CWTabBar, CWTab } from './cw_tabs';
 import { CWText } from './cw_text';
 import { CWPopoverMenu } from './cw_popover/cw_popover_menu';
 import { CWIconButton } from './cw_icon_button';
 import { isWindowMediumSmallInclusive } from './helpers';
+import { CWIcon } from './cw_icons/cw_icon';
 
 type SidebarItem = {
   label: string;
@@ -27,7 +29,7 @@ type SidebarComponents = [
 
 type ContentPageAttrs = {
   createdAt: moment.Moment;
-  title: string;
+  title: string | m.Vnode;
 
   // optional
   author?: m.Vnode;
@@ -35,10 +37,12 @@ type ContentPageAttrs = {
   body?: m.Vnode;
   comments?: Array<Comment<any>>;
   contentBodyLabel?: 'Snapshot' | 'Thread'; // proposals don't need a label because they're never tabbed
+  readOnly?: boolean;
   showSidebar?: boolean;
   sidebarComponents?: SidebarComponents;
   subBody?: m.Vnode;
   subHeader?: m.Vnode;
+  viewCount?: number;
 };
 
 export class CWContentPage implements m.ClassComponent<ContentPageAttrs> {
@@ -83,24 +87,34 @@ export class CWContentPage implements m.ClassComponent<ContentPageAttrs> {
       comments,
       contentBodyLabel,
       createdAt,
+      readOnly,
       showSidebar,
       sidebarComponents,
       subBody,
       subHeader,
       title,
+      viewCount,
     } = vnode.attrs;
 
     const mainBody = (
       <div class="main-body-container">
         <div class="header">
-          <CWText type="h3" fontWeight="semiBold">
-            {title}
-          </CWText>
+          {typeof title === 'string' ? (
+            <CWText type="h3" fontWeight="semiBold">
+              {title}
+            </CWText>
+          ) : (
+            title
+          )}
           <div class="header-info-row">
             {author}
             <CWText type="caption" className="header-text">
               published on {moment(createdAt).format('l')}
             </CWText>
+            <CWText type="caption" className="header-text">
+              {pluralize(viewCount, 'view')}
+            </CWText>
+            {readOnly && <CWIcon iconName="lock" iconSize="small" />}
             {actions && (
               <CWPopoverMenu
                 trigger={<CWIconButton iconName="dotsVertical" />}
