+++ packages/commonwealth/client/scripts/views/pages/view_thread/index.tsx
@@ -12,18 +12,36 @@ import { ProposalType } from 'common-common/src/types';
 import { getProposalUrlPath, idToProposal } from 'identifiers';
 import { slugify } from 'utils';
 import { notifyError } from 'controllers/app/notifications';
-import { Comment, Poll, Thread } from 'models';
+import {
+  Comment,
+  Poll,
+  Thread,
+  ThreadStage as ThreadStageType,
+  Topic,
+} from 'models';
+import { ContentType } from 'types';
 import { PageLoading } from 'views/pages/loading';
 import { PageNotFound } from 'views/pages/404';
+import TopicGateCheck from 'controllers/chain/ethereum/gatedTopic';
 import { modelFromServer as modelReactionCountFromServer } from 'controllers/server/reactionCounts';
-import { CWTabBar, CWTab } from '../../components/component_kit/cw_tabs';
-import {
-  getClasses,
-  isWindowMediumSmallInclusive,
-} from '../../components/component_kit/helpers';
 import { activeQuillEditorHasText } from './helpers';
 import { ThreadSidebar } from './thread_sidebar';
-import { ThreadBody } from './thread_body';
+import { CWContentPage } from '../../components/component_kit/cw_content_page';
+import { CWTextInput } from '../../components/component_kit/cw_text_input';
+import { ExternalLink, ThreadAuthor, ThreadStage } from './thread_components';
+import { CommentsTree } from '../../components/comments/comments_tree';
+import { SharePopover } from '../../components/share_popover';
+import { clearEditingLocalStorage } from '../../components/comments/helpers';
+import { confirmationModalWithText } from '../../modals/confirm_modal';
+import { EditCollaboratorsModal } from '../../modals/edit_collaborators_modal';
+import { ChangeTopicModal } from '../../modals/change_topic_modal';
+import { getThreadSubScriptionMenuItem } from '../discussions/discussion_row_menu';
+import { CollapsibleBodyText } from '../../components/collapsible_body_text';
+import { CreateComment } from '../../components/comments/create_comment';
+import { CWIcon } from '../../components/component_kit/cw_icons/cw_icon';
+import { CWText } from '../../components/component_kit/cw_text';
+import { ThreadReactionButton } from '../../components/reaction_button/thread_reaction_button';
+import { EditBody } from './edit_body';
 
 export type ThreadPrefetch = {
   [identifier: string]: {
@@ -42,19 +60,18 @@ class ViewThreadPage
     }>
 {
   private comments: Array<Comment<Thread>>;
+  private isEditingBody: boolean;
   private isGloballyEditing: boolean;
   private polls: Poll[];
   private prefetch: ThreadPrefetch;
-  private thread: Thread;
   private recentlyEdited: boolean;
-  private tabSelected: 'viewProposal' | 'viewSidebar';
+  private savedEdits: string;
+  private shouldRestoreEdits: boolean;
+  private thread: Thread;
   private threadFetched: boolean;
   private threadFetchFailed: boolean;
+  private title: string;
   private viewCount: number;
-
-  oninit() {
-    this.tabSelected = 'viewProposal';
-  }
 
   view(vnode) {
     const { identifier } = vnode.attrs;
@@ -140,15 +157,19 @@ class ViewThreadPage
       }
     }
 
-    const { thread } = this;
+    if (this.thread) {
+      this.title = this.thread.title;
+    }
 
-    if (threadRecentlyEdited) this.recentlyEdited = false;
+    if (threadRecentlyEdited) {
+      this.recentlyEdited = false;
+    }
 
-    if (identifier !== `${threadId}-${slugify(thread.title)}`) {
+    if (identifier !== `${threadId}-${slugify(this.thread.title)}`) {
       navigateToSubpage(
         getProposalUrlPath(
-          thread.slug,
-          `${threadId}-${slugify(thread.title)}`,
+          this.thread.slug,
+          `${threadId}-${slugify(this.thread.title)}`,
           true
         ),
         {},
@@ -158,17 +179,17 @@ class ViewThreadPage
 
     // load proposal
     if (!this.prefetch[threadIdAndType]['threadReactionsStarted']) {
-      app.threads.fetchReactionsCount([thread]).then(() => m.redraw);
+      app.threads.fetchReactionsCount([this.thread]).then(() => m.redraw);
       this.prefetch[threadIdAndType]['threadReactionsStarted'] = true;
     }
 
     // load comments
     if (!this.prefetch[threadIdAndType]['commentsStarted']) {
       app.comments
-        .refresh(thread, app.activeChainId())
+        .refresh(this.thread, app.activeChainId())
         .then(async () => {
           this.comments = app.comments
-            .getByProposal(thread)
+            .getByProposal(this.thread)
             .filter((c) => c.parentComment === null);
 
           // fetch reactions
@@ -181,7 +202,7 @@ class ViewThreadPage
             data: JSON.stringify({
               proposal_ids: [threadId],
               comment_ids: app.comments
-                .getByProposal(thread)
+                .getByProposal(this.thread)
                 .map((comment) => comment.id),
               active_address: app.user.activeAccount?.address,
             }),
@@ -222,14 +243,14 @@ class ViewThreadPage
 
     const updatedCommentsCallback = () => {
       this.comments = app.comments
-        .getByProposal(thread)
+        .getByProposal(this.thread)
         .filter((c) => c.parentComment === null);
       m.redraw();
     };
 
     // load polls
     if (!this.prefetch[threadIdAndType]['pollsStarted']) {
-      app.polls.fetchPolls(app.activeChainId(), thread.id).catch(() => {
+      app.polls.fetchPolls(app.activeChainId(), this.thread.id).catch(() => {
         notifyError('Failed to load comments');
         this.comments = [];
         m.redraw();
@@ -237,14 +258,14 @@ class ViewThreadPage
 
       this.prefetch[threadIdAndType]['pollsStarted'] = true;
     } else {
-      this.polls = app.polls.getByThreadId(thread.id);
+      this.polls = app.polls.getByThreadId(this.thread.id);
     }
 
     // load view count
     if (!this.prefetch[threadIdAndType]['viewCountStarted']) {
       $.post(`${app.serverUrl()}/viewCount`, {
         chain: app.activeChainId(),
-        object_id: thread.id,
+        object_id: this.thread.id,
       })
         .then((response) => {
           if (response.status !== 'Success') {
@@ -273,7 +294,7 @@ class ViewThreadPage
 
     // load profiles
     if (this.prefetch[threadIdAndType]['profilesStarted'] === undefined) {
-      app.profiles.getProfile(thread.authorChain, thread.author);
+      app.profiles.getProfile(this.thread.authorChain, this.thread.author);
 
       this.comments.forEach((comment) => {
         app.profiles.getProfile(comment.authorChain, comment.author);
@@ -295,18 +316,18 @@ class ViewThreadPage
 
     this.prefetch[threadIdAndType]['profilesFinished'] = true;
 
-    const commentCount = app.comments.nComments(thread);
+    const commentCount = app.comments.nComments(this.thread);
 
     // Original posters have full editorial control, while added collaborators
     // merely have access to the body and title
     const { activeAccount } = app.user;
 
     const isAuthor =
-      activeAccount?.address === thread.author &&
-      activeAccount?.chain.id === thread.authorChain;
+      activeAccount?.address === this.thread.author &&
+      activeAccount?.chain.id === this.thread.authorChain;
 
     const isEditor =
-      thread.collaborators?.filter((c) => {
+      this.thread.collaborators?.filter((c) => {
         return (
           c.address === activeAccount?.address &&
           c.chain === activeAccount?.chain.id
@@ -326,34 +347,14 @@ class ViewThreadPage
         chain: app.activeChainId(),
       });
 
-    const setIsGloballyEditing = (status: boolean) => {
-      this.isGloballyEditing = status;
-
-      if (status === false) {
-        this.recentlyEdited = true;
-      }
-
-      m.redraw();
-    };
-
     const showLinkedSnapshotOptions =
-      thread.snapshotProposal?.length > 0 ||
-      thread.chainEntities?.length > 0 ||
+      this.thread.snapshotProposal?.length > 0 ||
+      this.thread.chainEntities?.length > 0 ||
       isAuthor ||
       isAdminOrMod;
 
     const showLinkedThreadOptions =
-      thread.linkedThreads?.length > 0 || isAuthor || isAdminOrMod;
-
-    window.onresize = () => {
-      if (
-        isWindowMediumSmallInclusive(window.innerWidth) &&
-        this.tabSelected !== 'viewProposal'
-      ) {
-        this.tabSelected = 'viewProposal';
-        m.redraw();
-      }
-    };
+      this.thread.linkedThreads?.length > 0 || isAuthor || isAdminOrMod;
 
     const windowListener = (e) => {
       if (this.isGloballyEditing || activeQuillEditorHasText()) {
@@ -364,96 +365,280 @@ class ViewThreadPage
 
     window.addEventListener('beforeunload', windowListener);
 
-    const hasSidebar =
-      showLinkedSnapshotOptions ||
-      showLinkedThreadOptions ||
-      this.polls?.length > 0 ||
-      isAuthor;
+    const canComment =
+      app.user.activeAccount ||
+      (!isAdminOrMod && TopicGateCheck.isGatedTopic(this.thread?.topic?.name));
+
+    const setIsGloballyEditing = (status: boolean) => {
+      this.isGloballyEditing = status;
+
+      if (status === false) {
+        this.recentlyEdited = true;
+      }
+
+      m.redraw();
+    };
+
+    const setIsEditingBody = (status: boolean) => {
+      setIsGloballyEditing(status);
+      this.isEditingBody = status;
+    };
+
+    const reactionsAndReplyButtons = (
+      <div class="thread-footer-row">
+        <ThreadReactionButton thread={this.thread} />
+        <div class="comments-count">
+          <CWIcon iconName="feedback" iconSize="small" />
+          <CWText type="caption">{commentCount} Comments</CWText>
+        </div>
+      </div>
+    );
+
+    const hasEditPerms = isAuthor || isAdminOrMod || isEditor;
 
     return (
       <Sublayout
       //  title={headerTitle}
       >
-        <div class="ViewProposalPage">
-          <div
-            class={getClasses<{ hasSidebar?: boolean }>(
-              { hasSidebar },
-              'proposal-body-with-tabs'
-            )}
-          >
-            <CWTabBar>
-              <CWTab
-                label="Proposal"
-                onclick={() => {
-                  this.tabSelected = 'viewProposal';
+        <CWContentPage
+          contentBodyLabel="Thread"
+          showSidebar={
+            showLinkedSnapshotOptions ||
+            showLinkedThreadOptions ||
+            this.polls?.length > 0 ||
+            isAuthor
+          }
+          title={
+            this.isEditingBody ? (
+              <CWTextInput
+                oninput={(e) => {
+                  this.title = e.target.value;
                 }}
-                isSelected={this.tabSelected === 'viewProposal'}
+                value={this.title}
               />
-              <CWTab
-                label="Info & Results"
-                onclick={() => {
-                  this.tabSelected = 'viewSidebar';
-                }}
-                isSelected={this.tabSelected === 'viewSidebar'}
-              />
-            </CWTabBar>
-            {this.tabSelected === 'viewProposal' && (
-              <ThreadBody
-                commentCount={commentCount}
-                comments={this.comments}
-                updatedCommentsCallback={updatedCommentsCallback}
-                setIsGloballyEditing={setIsGloballyEditing}
-                isAdminOrMod={isAdminOrMod}
-                isAuthor={isAuthor}
-                isEditor={isEditor}
-                isGloballyEditing={this.isGloballyEditing}
-                proposal={thread}
-                viewCount={this.viewCount}
-              />
-            )}
-            {hasSidebar && this.tabSelected === 'viewSidebar' && (
-              <ThreadSidebar
-                isAdmin={isAdmin}
-                isAdminOrMod={isAdminOrMod}
-                isAuthor={isAuthor}
-                polls={this.polls}
-                thread={thread}
-                showLinkedSnapshotOptions={showLinkedSnapshotOptions}
-                showLinkedThreadOptions={showLinkedThreadOptions}
-              />
-            )}
-          </div>
-          <div
-            class={getClasses<{ hasSidebar?: boolean }>(
-              { hasSidebar },
-              'proposal-body'
-            )}
-          >
-            <ThreadBody
-              commentCount={commentCount}
+            ) : (
+              this.thread.title
+            )
+          }
+          author={<ThreadAuthor thread={this.thread} />}
+          createdAt={this.thread.createdAt}
+          viewCount={this.viewCount}
+          readOnly={this.thread.readOnly}
+          subHeader={
+            <>
+              {this.thread.stage !== ThreadStageType.Discussion && (
+                <ThreadStage thread={this.thread} />
+              )}
+              <SharePopover />
+              {!!this.thread.url && <ExternalLink thread={this.thread} />}
+            </>
+          }
+          actions={
+            app.isLoggedIn() &&
+            hasEditPerms &&
+            !this.isGloballyEditing && [
+              ...(hasEditPerms && !this.thread.readOnly
+                ? [
+                    {
+                      label: 'Edit',
+                      iconLeft: 'write',
+                      onclick: async (e) => {
+                        e.preventDefault();
+                        this.savedEdits = localStorage.getItem(
+                          `${app.activeChainId()}-edit-thread-${
+                            this.thread.id
+                          }-storedText`
+                        );
+
+                        if (this.savedEdits) {
+                          clearEditingLocalStorage(
+                            this.thread.id,
+                            ContentType.Thread
+                          );
+                          this.shouldRestoreEdits =
+                            await confirmationModalWithText(
+                              'Previous changes found. Restore edits?',
+                              'Yes',
+                              'No'
+                            )();
+                        }
+
+                        setIsEditingBody(true);
+                      },
+                    },
+                  ]
+                : []),
+              ...(isAuthor
+                ? [
+                    {
+                      label: 'Edit collaborators',
+                      iconLeft: 'write',
+                      onclick: async (e) => {
+                        e.preventDefault();
+                        app.modals.create({
+                          modal: EditCollaboratorsModal,
+                          data: {
+                            thread: this.thread,
+                          },
+                        });
+                      },
+                    },
+                  ]
+                : []),
+              ...(isAdminOrMod
+                ? [
+                    {
+                      label: 'Change topic',
+                      iconLeft: 'write',
+                      onclick: (e) => {
+                        e.preventDefault();
+                        app.modals.create({
+                          modal: ChangeTopicModal,
+                          data: {
+                            onChangeHandler: (topic: Topic) => {
+                              this.thread.topic = topic;
+                              m.redraw();
+                            },
+                            thread: this.thread,
+                          },
+                        });
+                      },
+                    },
+                  ]
+                : []),
+              ...(isAuthor || isAdminOrMod || app.user.isSiteAdmin
+                ? [
+                    {
+                      label: 'Delete',
+                      iconLeft: 'trash',
+                      onclick: async (e) => {
+                        e.preventDefault();
+
+                        const confirmed = await confirmationModalWithText(
+                          'Delete this entire thread?'
+                        )();
+
+                        if (!confirmed) return;
+
+                        app.threads.delete(this.thread).then(() => {
+                          navigateToSubpage('/discussions');
+                        });
+                      },
+                    },
+                  ]
+                : []),
+              ...(isAuthor || isAdminOrMod
+                ? [
+                    {
+                      label: this.thread.readOnly
+                        ? 'Unlock thread'
+                        : 'Lock thread',
+                      iconLeft: 'lock',
+                      onclick: (e) => {
+                        e.preventDefault();
+                        app.threads
+                          .setPrivacy({
+                            threadId: this.thread.id,
+                            readOnly: !this.thread.readOnly,
+                          })
+                          .then(() => {
+                            setIsEditingBody(false);
+                            m.redraw();
+                          });
+                      },
+                    },
+                  ]
+                : []),
+              ...((isAuthor || isAdminOrMod) &&
+              !!app.chain?.meta.snapshot.length
+                ? [
+                    {
+                      label: 'Snapshot proposal from thread',
+                      iconLeft: 'democraticProposal',
+                      onclick: () => {
+                        const snapshotSpaces = app.chain.meta.snapshot;
+
+                        if (snapshotSpaces.length > 1) {
+                          navigateToSubpage('/multiple-snapshots', {
+                            action: 'create-from-thread',
+                            thread: this.thread,
+                          });
+                        } else {
+                          navigateToSubpage(`/snapshot/${snapshotSpaces}`);
+                        }
+                      },
+                    },
+                  ]
+                : []),
+              ...(isAuthor || isAdminOrMod
+                ? [
+                    { type: 'divider' },
+                    getThreadSubScriptionMenuItem(this.thread),
+                  ]
+                : []),
+            ]
+          }
+          body={
+            <div class="thread-content">
+              {this.isEditingBody ? (
+                <>
+                  {reactionsAndReplyButtons}
+                  <EditBody
+                    thread={this.thread}
+                    savedEdits={this.savedEdits}
+                    shouldRestoreEdits={this.shouldRestoreEdits}
+                    setIsEditing={setIsEditingBody}
+                    title={this.title}
+                  />
+                </>
+              ) : (
+                <>
+                  <CollapsibleBodyText item={this.thread} />
+                  {this.thread.readOnly ? (
+                    <CWText type="h5" className="callout-text">
+                      Commenting is disabled because this post has been locked.
+                    </CWText>
+                  ) : !this.isGloballyEditing && canComment ? (
+                    <>
+                      {reactionsAndReplyButtons}
+                      <CreateComment
+                        updatedCommentsCallback={updatedCommentsCallback}
+                        setIsGloballyEditing={setIsGloballyEditing}
+                        isGloballyEditing={this.isGloballyEditing}
+                        parentComment={null}
+                        rootProposal={this.thread}
+                      />
+                    </>
+                  ) : null}
+                </>
+              )}
+            </div>
+          }
+          comments={
+            <CommentsTree
               comments={this.comments}
-              updatedCommentsCallback={updatedCommentsCallback}
+              proposal={this.thread}
               setIsGloballyEditing={setIsGloballyEditing}
-              isAdminOrMod={isAdminOrMod}
-              isAuthor={isAuthor}
-              isEditor={isEditor}
-              isGloballyEditing={this.isGloballyEditing}
-              proposal={thread}
-              viewCount={this.viewCount}
+              updatedCommentsCallback={updatedCommentsCallback}
             />
-            {hasSidebar && (
-              <ThreadSidebar
-                isAdmin={isAdmin}
-                isAdminOrMod={isAdminOrMod}
-                isAuthor={isAuthor}
-                polls={this.polls}
-                thread={thread}
-                showLinkedSnapshotOptions={showLinkedSnapshotOptions}
-                showLinkedThreadOptions={showLinkedThreadOptions}
-              />
-            )}
-          </div>
-        </div>
+          }
+          sidebarComponents={[
+            {
+              label: 'Info',
+              item: (
+                <ThreadSidebar
+                  isAdmin={isAdmin}
+                  isAdminOrMod={isAdminOrMod}
+                  isAuthor={isAuthor}
+                  polls={this.polls}
+                  thread={this.thread}
+                  showLinkedSnapshotOptions={showLinkedSnapshotOptions}
+                  showLinkedThreadOptions={showLinkedThreadOptions}
+                />
+              ),
+            },
+          ]}
+        />
       </Sublayout>
     );
   }
